+++ src/memory_bank_controllers/no_mbc.cc
@@ -3,7 +3,7 @@
 #include "no_mbc.h"
 
 
-bool NoMBC::read(uint16_t addr, uint16_t &mapped_addr, bool &rom_read) {
+bool NoMBC::read(uint16_t addr, uint32_t &mapped_addr, bool &rom_read) {
     if (addr >= 0x0000 && addr < 0x8000) {
         mapped_addr = addr;
         rom_read = true;
@@ -17,7 +17,7 @@ bool NoMBC::read(uint16_t addr, uint16_t &mapped_addr, bool &rom_read) {
     return false;
 }
 
-bool NoMBC::write(uint16_t addr, uint8_t data, uint16_t &mapped_addr) {
+bool NoMBC::write(uint16_t addr, uint8_t data, uint32_t &mapped_addr) {
     if (addr >= 0xA000 && addr < 0xC000) {
         mapped_addr = addr;
         return true;
